requirements.txt
flask
Flask-SocketIO
weaviate-client==4.*
python-dotenv
pypdf
pdf2image
pytesseract
openai
supabase
